app/requirements.txt
Flask
openai
faiss-cpu
sentence-transformers